+++ src/browserify.index.js
@@ -1,3 +1,4 @@
+import 'babel-polyfill';
 import * as RxDB from './index.js';
 
 RxDB.plugin(require('pouchdb-adapter-idb'));